moving recover, adding confirmation for deletion
